dosta malih promena

+++ Command line interpreter/Command.cpp
@@ -43,7 +43,7 @@ string Command::toString() {
 
 //checks if the last argument is a file
 bool Command::isArgFile(string arg) {
-	size_t argSize = arg.size() - 1;
+	int argSize = arg.size() - 1;
 	if (argSize < 4) return false;
 
 	string end = arg.substr(argSize - 3, argSize);
